img_capture.py: Ensure to only send msg 4 to app when exactly one Image is downloaded and renamed from the camera. Because of the implementation to delete all files from the camera before and after taking the photo, we ensure to have only the latest image on the host system. With this we prevent printing 2 pictures when e.g. the picture before was not downloaded or deleted correctly. This implementation is not designed to process multiple pictures, only works as intended with one picture only on the camera.
print.py: improved error handling, make sure to use relative file paths, removed changing directory's

New Error Codes and success code:
110 general error in prin.py
112 printer not found
113 file not found (print.py)
114 permission denied (print.py)
115 error copy file

204 print.py successfully finished

+++ Python/print.py
@@ -4,6 +4,7 @@ import socket
 import cups
 import os
 import config
+import time
 
 def send_message_to_app(message):
     try:
@@ -15,76 +16,68 @@ def send_message_to_app(message):
 
 def print_image(printer_name, image_path):
     print("Connect to the CUPS printing Server")
-    # Connect to CUPS
     conn = cups.Connection()
-    # Get a list of all printers
     printers = conn.getPrinters()
 
-    # Check if specified printer is available
     if printer_name not in printers:
         print(f"Printer {printer_name} not found. Available printers:")
-        send_message_to_app("110")
+        send_message_to_app("112")
         for printer in printers:
             print(f"{printer}\n")
         return
 
-    print(f"Sucessfully Connected to the print Server {printer_name}")
+    print(f"Successfully Connected to the print Server {printer_name}")
 
     if config.DEBUG_MODE == 0:
-        # Print the image
+        if not os.path.exists(image_path):
+            print(f"Error: File {image_path} not found.")
+            send_message_to_app("113")
+            return
+
         try:
             print_job_id = conn.printFile(printer_name, image_path, "Photo Print", {})
-            print(
-                f"Print job submitted to Printer. Job ID: {print_job_id}\nStart printing {image_path} on {printer_name}")
+            print(f"Print job submitted. Job ID: {print_job_id} - {image_path} on {printer_name}")
         except Exception as e:
-            print(f"Error while creating and sending print job.")
-            send_message_to_app("110")
+            print(f"Error in print job: {e}")
+            send_message_to_app("119")
     else:
         print(f"DEBUG MODE: Simulate Print file {image_path} on {printer_name}. \nDEBUG MODE: Skip 45 sec waiting time...")
 
 def copy_file(source_path, destination_path):
+    if not os.path.exists(source_path):
+        print(f"Error: The file {source_path} does not exist.")
+        send_message_to_app("113")
+        return False
+
     try:
-        # Copy the source file to the destination path
         shutil.move(source_path, destination_path)
         print(f"Successfully copied {source_path} to {destination_path}")
-    except FileNotFoundError:
-        print(f"CWD {os.getcwd()}")
-        print(f"Error: The file {source_path} does not exist.")
-        send_message_to_app("103")
+        return True
     except PermissionError:
-        print(f"Error: Permission denied while copying {source_path}.")
-        send_message_to_app("104")
+        print(f"Error: Permission denied for {source_path}.")
+        send_message_to_app("114")
     except Exception as e:
-        print(f"An unexpected error occurred while copying {source_path}: {str(e)}")
-        send_message_to_app("100")
+        print(f"Error copying {source_path}: {e}")
+        send_message_to_app("115")
 
+    return False
 
 def move_image():
-    pic_dir = os.path.join(config.PICTURE_SAVE_DIRECTORY, datetime.datetime.now().strftime("%Y-%m-%d"))
+    pic_dir = os.path.abspath(os.path.join(config.PICTURE_SAVE_DIRECTORY, datetime.datetime.now().strftime("%Y-%m-%d")))
     printer_name = config.PRINTER_NAME
-    cwd = os.getcwd()
-    os.chdir(pic_dir)
-    cwd_tmp = os.getcwd()
-    print(f"Change directory to {pic_dir}")
 
-    for filename in os.listdir():
-        picture_name = filename
-        # Construct the full source and destination paths
-        source_picture_path = os.path.join(cwd_tmp, picture_name)
-        # swap print and pic dir in the source picture path and remove ".." before if it is there in config eg: ../images/print/ -> /images/print/ to build destination path for picture
-        destination_picture_path = os.path.join(cwd_tmp.replace(f"{config.PICTURE_SAVE_DIRECTORY.replace('..','')}",f"{config.PRINT_DIR.replace('..','')}"))
-        # copy picture from source path to the desination path
-        copy_file(source_picture_path, destination_picture_path)
-        print_image(printer_name, destination_picture_path)
-        os.remove(destination_picture_path)
-        print(f'Successfully removed {picture_name} from Disk.')
+    for filename in os.listdir(pic_dir):
+        source_path = os.path.join(pic_dir, filename)
+        destination_path = os.path.join(config.PRINT_DIR, filename)
 
-    os.chdir(cwd)
-    print(f"Change directory back to {cwd}")
+        if copy_file(source_path, destination_path):
+            print_image(printer_name, destination_path)
+            os.remove(destination_path)
+            print(f'Removed {filename} after creating and sending print Job.')
+            send_message_to_app("204")
 
 if __name__ == '__main__':
-    print("print.py is now running.")
-    print(f"preparing print Job")
+    time.sleep(8)
+    print("Starting print process.")
     move_image()
-    print(f"Printing now up to 45 seconds ...")
-    print("print.py is now finished successfully")
+    print("Print process completed.")
